+++ CONTRIBUTING.md
@@ -180,36 +180,3 @@ To add a new banner, follow these steps:
 4. Upload both files to `/public/promo-banners/`.
 5. Fill out `imageUrl`, `imageUrlMobile` and optionally `redirecturl` inside [`src/app/layout.tsx`](https://github.com/DeXter-on-Radix/website/blob/main/src/app/layout.tsx#L15-L20).
 
-## Hydration Error Handling
-
-**Problem**
-
-State variables get cached, leading to unpredictable initial renders of components (e.g., initialization based on cookies). This violates React/NextJS patterns and triggers hydration errors.
-
-**Handling Hydration Errors**
-
-1. **No Error**: If no hydration error occurs, no action is needed.
-2. **Error Detected**: Identify the component causing the error.
-3. **Fixing the Component**: Add the following code to the problematic component:
-
-```tsx
-import { useHydrationErrorFix } from "hooks";
-
-function ComponentWithHydrationError() {
-  const isClient = useHydrationErrorFix();
-
-  // Additional code like useEffect or other hooks go here!
-  // ...
-
-  if (!isClient) return <></>;
-
-  return (
-    /* Your component JSX */
-  );
-}
-```
-
-**Common Causes of Hydration Errors**
-
-1. **Radix Connect Button**: Caches logged-in users. Components relying on login status will render differently based on the user's login state.
-2. **Language Detection**: Cached in cookies. Components will render differently for users from various regions based on their browser-detected language.
