+++ src/app/components/OrderInput.tsx
@@ -8,12 +8,7 @@ import {
   truncateWithPrecision,
 } from "../utils";
 
-import {
-  useAppDispatch,
-  useAppSelector,
-  useTranslations,
-  useHydrationErrorFix,
-} from "hooks";
+import { useAppDispatch, useAppSelector, useTranslations } from "hooks";
 import { fetchBalances } from "state/pairSelectorSlice";
 import {
   OrderSide,
@@ -360,7 +355,6 @@ function PostOnlyCheckbox() {
 }
 
 function SubmitButton() {
-  const isClient = useHydrationErrorFix(); // to fix HydrationError
   const t = useTranslations();
   const dispatch = useAppDispatch();
   const {
@@ -390,9 +384,6 @@ function SubmitButton() {
         .replaceAll("<$ORDER_TYPE>", t(type))
         .replaceAll("<$SIDE>", t(side))
         .replaceAll("<$TOKEN_SYMBOL>", token1.symbol);
-
-  // Fix HydrationError
-  if (!isClient) return <></>;
 
   return (
     <button
