+++ src/app/rewards/page.tsx
@@ -1,12 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import {
-  useAppDispatch,
-  useAppSelector,
-  useTranslations,
-  useHydrationErrorFix,
-} from "hooks";
+import { useAppDispatch, useAppSelector, useTranslations } from "hooks";
 import {
   fetchAddresses,
   fetchReciepts,
@@ -97,7 +92,6 @@ function HeaderComponent() {
 }
 
 function RewardsCard() {
-  const isClient = useHydrationErrorFix(); // to fix HydrationError
   const dispatch = useAppDispatch();
   const { isConnected, selectedAccount } = useAppSelector(
     (state) => state.radix
@@ -133,9 +127,6 @@ function RewardsCard() {
       loadRewards();
     }
   }, [dispatch, isConnected, account, pairsList]);
-
-  // Fix HydrationError
-  if (!isClient) return <></>;
 
   return (
     <div className="max-w-[400px] sm:max-w-[600px] px-4 py-4 sm:px-12 sm:py-8 m-auto mt-2 sm:mt-14 mb-28 bg-[#191B1D] rounded-xl max-[450px]:mx-5">
@@ -219,7 +210,6 @@ function RewardsOverview() {
 }
 
 function ClaimButton() {
-  const isClient = useHydrationErrorFix(); // to fix HydrationError
   const t = useTranslations();
   const dispatch = useAppDispatch();
   const { isConnected } = useAppSelector((state) => state.radix);
@@ -228,9 +218,6 @@ function ClaimButton() {
   );
   const userHasRewards = getUserHasRewards(rewardData);
   const disabled = !isConnected || !userHasRewards;
-
-  // Fix HydrationError
-  if (!isClient) return <></>;
 
   return (
     <button
@@ -268,7 +255,6 @@ function ClaimButton() {
 
 function RewardsDetails() {
   const [isOpen, setIsOpen] = useState(true);
-  const isClient = useHydrationErrorFix(); // to fix HydrationError
   const { isConnected } = useAppSelector((state) => state.radix);
   const { rewardData, tokensList } = useAppSelector(
     (state) => state.rewardSlice
@@ -281,7 +267,7 @@ function RewardsDetails() {
     }
   }, [isConnected]);
 
-  if (!isConnected || !userHasRewards || !isClient) {
+  if (!isConnected || !userHasRewards) {
     return <></>;
   }
